refactor: 패키지를 역할에 따라 구조화
