+++ src/main/kotlin/lotto/domain/model/purchaseInfo/quantity/LottoQuantity.kt
@@ -1,4 +1,4 @@
-package lotto.domain.valueobject
+package lotto.domain.model.purchaseInfo.quantity
 
 @JvmInline
 value class LottoQuantity(
