+++ src/test/kotlin/lottotest/domain/model/purchaseInfo/quantity/LottoQuantityTest.kt
@@ -1,6 +1,6 @@
-package lottotest.domain.valueobject
+package lottotest.domain.model.purchaseInfo.quantity
 
-import lotto.domain.valueobject.LottoQuantity
+import lotto.domain.model.purchaseInfo.quantity.LottoQuantity
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
