+++ src/test/kotlin/lottotest/domain/model/purchaseInfo/quantity/ManualLottoQuantityTest.kt
@@ -1,7 +1,7 @@
-package lottotest.domain.valueobject
+package lottotest.domain.model.purchaseInfo.quantity
 
-import lotto.domain.valueobject.LottoQuantity
-import lotto.domain.valueobject.ManualLottoQuantity
+import lotto.domain.model.purchaseInfo.quantity.LottoQuantity
+import lotto.domain.model.purchaseInfo.quantity.ManualLottoQuantity
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
