+++ src/test/kotlin/lottotest/domain/model/purchaseInfo/ticket/AutoLottoTicketTest.kt
@@ -1,7 +1,7 @@
-package lottotest.domain.model.lottoticket
+package lottotest.domain.model.purchaseInfo.ticket
 
-import lotto.domain.model.lottoticket.AutoLottoTicket
-import lotto.domain.model.lottoticket.LottoTicket
+import lotto.domain.model.purchaseInfo.ticket.AutoLottoTicket
+import lotto.domain.model.purchaseInfo.ticket.LottoTicket
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
